requirements.txt
email_validator
dnspython
